Bot.py
This bot will help with the whole foods market checkout to automatically log you in and go to check out and if no options are available it will refresh page till there is an option. Also, this has some tactics to make the Recaptcha not appear like randomly writing at a different pace, changing window size, and changing the agent user every time it reboots. Also if there is an error of some sort it will reboot itself and close down the browser and start a new one up from the beginning. This is still a work in progress but it will be updated more.