Redesign
Fixed some issues with the typing speed. Made the code more readable and clean. Also added fresh market and an option to choose if you want fresh market or whole foods market.  I also made it so that it can save your email and password so that you won't have to constantly write your email and password and just need to do it once. Also, there is an option to change the email and password in case you messed up or want to change it.